Backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
pandas
python-dotenv
